changing video and stream into queue
