+++ main.py
@@ -2,12 +2,14 @@ from tflite_loader import model_loader
 # import tflite_runtime.interpreter as mytflite
 import argparse
 import numpy as np
-from data_loader import data_loader
+from data_loader2 import data_loader2
 import time 
 import os
 import cv2
 import tensorflow as tf
+from queue import Queue
 """
+Average time is: 142.49318310139603
 I should write a test for all possible inputs
 for example when I change a thing I should run these both:
 
@@ -23,7 +25,9 @@ def parsing():
     parser = argparse.ArgumentParser(description='',
                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('--img_path', help='Path for image file', type=str, default=None)
+    parser.add_argument('--img_folder', help='Path for image folder', type=str, default=None)
     parser.add_argument('--vid_path', help='Path for video file', type=str, default=None)
+    parser.add_argument('--stream', help='Stream from device', type=int, default=None)
     parser.add_argument('--model_path', help='Model path file', type=str, required=True)
     parser.add_argument('--device', help='Device can be cuda or cpu or None', type=str, default=None)
     parser.add_argument('--delegate_path', help='File path of ArmNN delegate file', type=str, default=None)
@@ -33,6 +37,20 @@ def parsing():
 
     return args
 
+def queue_reader(loader, vid):
+    times = []
+    loader.load_vid(vid)
+    while True:
+        data = data_queue.get()
+        s = time.time()
+        out = model.inference(data)
+        e = time.time()
+        times.append(e-s)
+        print(f" Inference time is: {e-s}")
+        if not loader.ret:
+            break  # Break the loop if the video has ended
+
+    print(f" Average time is: {1 / np.mean(times)}")
 if __name__ == '__main__':
     args = parsing()
     # delegate_path = args.delegate_path
@@ -48,86 +66,28 @@ if __name__ == '__main__':
         os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
     model = model_loader(args)
     print("Model initiated")
-    loader = data_loader(img_shape=model.input_shape, model_shape=model.model_shape)
+    data_queue = Queue()
+    loader = data_loader2(img_shape=model.input_shape, model_shape=model.model_shape, data_queue=data_queue)
+
     print("Data loader initiated")
     if args.img_path is not None:
         imgs = loader.load_img(args.img_path)
-    elif args.vid_path is not None:
-        # imgs = loader.load_vid(args.vid_path)
-        video_capture = cv2.VideoCapture(args.vid_path)
-        if not video_capture.isOpened():
-            print("Error: Could not open the video file.")
-            exit()
-
-        times=[]
-        while True:
-            ret, frame = video_capture.read()
-            if not ret:
-                break  # Break the loop if the video has ended
-
-            img_np = cv2.resize(frame, (model.input_shape, model.input_shape))
-            if img_np.shape[0] != model.model_shape[1]:
-                img_np = img_np.transpose(2, 0, 1)
-            img_tf = tf.convert_to_tensor(img_np, dtype=tf.float32)
-            img_tf = tf.expand_dims(img_tf , axis=0)
+        times = []
+        for img in imgs:
             s = time.time()
-            out = model.inference(img_tf)
+            out = model.inference(img)
             e = time.time()
             times.append(e-s)
-            # print(f" Inference time is: {e-s}")
+        
+        print(f"output is recived {out} \n shape is: {out.shape},\
+            argmax: {np.argmax(out, axis=1)}\n \
+                average time is: {np.mean(times)} average frame rate is: {1 / np.mean(times)}")
 
-            # Exit the loop if 'q' key is pressed
-            if cv2.waitKey(1) & 0xFF == ord('q'):
-                print(f" Average time is: {1 / np.mean(times)}")
-                break
-
-        # Release the video capture object and close any open windows
-        video_capture.release()
-        print(f" Average time is: {1 / np.mean(times)}")
+    elif None: 
+        pass
+    elif args.vid_path is not None:
+        queue_reader(loader, args.vid_path)
         exit()
     else:
-        video_capture = cv2.VideoCapture(0)
-        if not video_capture.isOpened():
-            print("Error: Could not open the video file.")
-            exit()
-
-        times=[]
-        while True:
-            ret, frame = video_capture.read()
-            if not ret:
-                break  # Break the loop if the video has ended
-
-            img_np = cv2.resize(frame, (model.input_shape, model.input_shape))
-            if img_np.shape[0] != model.model_shape[1]:
-                img_np = img_np.transpose(2, 0, 1)
-            img_tf = tf.convert_to_tensor(img_np, dtype=tf.float32)
-            img_tf = tf.expand_dims(img_tf , axis=0)
-            s = time.time()
-            out = model.inference(img_tf)
-            e = time.time()
-            times.append(e-s)
-            print(f" Inference time is: {e-s}")
-
-            # Exit the loop if 'q' key is pressed
-            if cv2.waitKey(1) & 0xFF == ord('q'):
-                print(f" Average time is: {1 / np.mean(times)}")
-                break
-
-        # Release the video capture object and close any open windows
-        video_capture.release()
-        print(f" Average time is: {1 / np.mean(times)}")
+        queue_reader(loader, args.stream)
         exit()
-
-
-    print("Data loaded")
-    
-    times = []
-    for img in imgs:
-        s = time.time()
-        out = model.inference(img)
-        e = time.time()
-        times.append(e-s)
-    
-    print(f"output is recived {out} \n shape is: {out.shape},\
-           argmax: {np.argmax(out, axis=1)}\n \
-            average time is: {np.mean(times)} average frame rate is: {1 / np.mean(times)}")
